+++ src/Components/PriceRange/utils/getValue.ts
@@ -1,0 +1,5 @@
+import { CustomRange } from "Components/PriceRange/constants"
+
+export const getValue = (value: CustomRange[number]) => {
+  return value === "*" || value === 0 ? "" : value
+}
